requirements.txt
fastapi
uvicorn
pydantic
httpx
firecrawl-py
langchain
langgraph
langchain-google-genai
sqlalchemy
pymysql
beautifulsoup4
google-api-python-client
